+++ Nos/Service/RelayService.swift
@@ -10,15 +10,17 @@ import Starscream
 import CoreData
 
 final class RelayService: ObservableObject {
-    private var requestFilterSet = Set<Filter>()
-    
-    private var sockets = [WebSocket]()
-    
     private var persistenceController: PersistenceController
+    private var requestFilterSet = Set<Filter>()
+    private var sockets = [WebSocket]()
+    private var timer: Timer?
     
     init(persistenceController: PersistenceController) {
         self.persistenceController = persistenceController
         openSocketsForRelays()
+        
+        let pubSel = #selector(publishFailedEvents)
+        timer = Timer.scheduledTimer(timeInterval: 120, target: self, selector: pubSel, userInfo: nil, repeats: true)
     }
     
     var allRelayAddresses: [String] {
@@ -118,16 +120,13 @@ extension RelayService {
 
         // Ignore redundant requests
         guard !requestFilterSet.contains(filter) else {
-            print("📡Request already open. Ignoring. \(requestFilterSet.count) outstanding requests")
+            print("Request with identical filter already open. Ignoring. \(requestFilterSet.count) filters in use.")
             let foundFilter = requestFilterSet.first(where: { $0 == filter })
             return foundFilter!.subscriptionId
         }
 
         requestFilterSet.insert(filter)
-        print("📡\(requestFilterSet.count) outstanding requests")
-        for request in requestFilterSet {
-            print("📡: \(request.dictionary)")
-        }
+        print("\(requestFilterSet.count) filters in use.")
 
         let subId = UUID().uuidString
         sockets.forEach { requestEvents(from: $0, subId: subId, filter: filter) }
@@ -150,34 +149,66 @@ extension RelayService {
     
     func parseEvent(_ responseArray: [Any]) {
         guard responseArray.count >= 3 else {
-            print("got invalid EVENT response: \(responseArray)")
+            print("Error: invalid EVENT response: \(responseArray)")
             return
         }
         
         guard let eventJSON = responseArray[2] as? [String: Any] else {
-            print("got invalid EVENT JSON: \(responseArray)")
+            print("Error: invalid EVENT JSON: \(responseArray)")
             return
         }
         
         do {
             _ = try EventProcessor.parse(jsonObject: eventJSON, in: persistenceController.container.viewContext)
         } catch {
-            print("error parsing event from relay: \(responseArray)")
+            print("Error: parsing event from relay: \(responseArray)")
         }
     }
     
-    func parseOK(_ responseArray: [Any]) {
+    func parseOK(_ responseArray: [Any], _ socket: WebSocket) {
         guard responseArray.count > 2 else {
             return
         }
         
-        if let result = responseArray[2] as? Bool, let subId = responseArray[1] as? String {
-            let resultString = result ? "sent succesfully" : "failed"
-            print("\(subId) has \(resultString)")
+        if let success = responseArray[2] as? Bool,
+            let eventId = responseArray[1] as? String,
+            let socketUrl = socket.request.url?.absoluteString {
+            let objectContext = persistenceController.container.viewContext
+            
+            if let event = Event.find(by: eventId, context: objectContext) {
+                let relay = Relay.findOrCreate(by: socketUrl, context: objectContext)
+
+                if success {
+                    print("\(eventId) has sent successfully to \(socketUrl)")
+                    if let pubRelays = event.publishedTo?.mutableCopy() as? NSMutableSet {
+                        pubRelays.add(relay)
+                        event.publishedTo = pubRelays
+                        print("Tracked publish to relay: \(socketUrl)")
+                    }
+                } else {
+                    // This will be picked up later in publishFailedEvents
+                    if responseArray.count > 2, let message = responseArray[3] as? String {
+                        // Mark duplicates or replaces as done on our end
+                        if message.contains("replaced:") || message.contains("duplicate:") {
+                            if let pubRelays = event.publishedTo?.mutableCopy() as? NSMutableSet {
+                                pubRelays.add(relay)
+                                event.publishedTo = pubRelays
+                                print("Tracked publish to relay: \(socketUrl)")
+                            }
+                        } else {
+                            print("\(eventId) has been rejected. Given reason: \(message)")
+                        }
+                    } else {
+                        print("\(eventId) has been rejected. No given reason.")
+                    }
+                }
+            } else {
+                print("Error: got OK for missing Event: \(eventId)")
+            }
         }
     }
     
-    func parseResponse(_ response: String) {
+    func parseResponse(_ response: String, _ socket: WebSocket) {
         do {
             guard let responseData = response.data(using: .utf8) else {
                 throw EventError.utf8Encoding
@@ -185,7 +216,7 @@ extension RelayService {
             let jsonResponse = try JSONSerialization.jsonObject(with: responseData)
             guard let responseArray = jsonResponse as? [Any],
                 let responseType = responseArray.first as? String else {
-                print("got unparseable response: \(response)")
+                print("Error: got unparseable response: \(response)")
                 return
             }
             switch responseType {
@@ -196,7 +227,7 @@ extension RelayService {
             case "EOSE":
                 parseEOSE(responseArray)
             case "OK":
-                parseOK(responseArray)
+                parseOK(responseArray, socket)
             default:
                 print("got unknown response type: \(response)")
             }
@@ -210,6 +241,8 @@ extension RelayService {
 extension RelayService {
     func publish(from client: WebSocketClient, event: Event) {
         do {
+            // Keep track of this so if it fails we can retry N times
+            event.sendAttempts += 1
             let request: [Any] = ["EVENT", event.jsonRepresentation!]
             let requestData = try JSONSerialization.data(withJSONObject: request)
             let requestString = String(data: requestData, encoding: .utf8)!
@@ -224,32 +257,62 @@ extension RelayService {
         openSocketsForRelays()
         sockets.forEach { publish(from: $0, event: event) }
     }
+    
+    @objc func publishFailedEvents() {
+        openSocketsForRelays()
+
+        let objectContext = persistenceController.container.viewContext
+        let userSentEvents = Event.allByUser(context: objectContext)
+        let relays = Relay.all(context: objectContext)
+        
+        // Only attempt to resend a user-created Event to Relays that were available at the time of publication
+        // This stops an Event from being sent to Relays that were added after the Event was sent
+        for event in userSentEvents {
+            let availableRelays = relays.filter { $0.createdAt! < event.createdAt! }
+            let publishedRelays: [Relay] = event.publishedTo?.allObjects as? [Relay] ?? []
+            let missedRelays: [Relay] = availableRelays.filter { !publishedRelays.contains($0) }
+            
+            print("\(missedRelays.count) missing a published event.")
+            for missedRelay in missedRelays {
+                guard let missedAddress = missedRelay.address else { continue }
+                if let index = sockets.firstIndex(where: { $0.request.url!.absoluteString == missedAddress }) {
+                    // Publish again to this socket
+                    print("Republishing \(event.identifier!) on \(missedAddress)")
+                    publish(from: sockets[index], event: event)
+                }
+            }
+        }
+    }
 }
 
 // MARK: WebSocketDelegate
 extension RelayService: WebSocketDelegate {
     func didReceive(event: WebSocketEvent, client: WebSocketClient) {
+        guard let socket = client as? WebSocket else {
+            return
+        }
+        
         switch event {
         case .connected(let headers):
             print("websocket is connected: \(headers)")
         case .disconnected(let reason, let code):
-            if let socket = client as? WebSocket, let index = sockets.firstIndex(where: { $0 === socket }) {
+            if let index = sockets.firstIndex(where: { $0 === socket }) {
                 sockets.remove(at: index)
             }
             print("websocket is disconnected: \(reason) with code: \(code)")
         case .text(let string):
-            parseResponse(string)
-            print("Received text: \(string)")
+            parseResponse(string, socket)
+            print("Received text (\(socket.request.url?.absoluteString ?? "unknown")): \(string)")
         case .binary(let data):
             print("Received data: \(data.count)")
         case .ping, .pong, .viabilityChanged, .reconnectSuggested:
             break
         case .cancelled:
-            if let socket = client as? WebSocket, let index = sockets.firstIndex(where: { $0 === socket }) {
+            if let index = sockets.firstIndex(where: { $0 === socket }) {
                 sockets.remove(at: index)
             }
         case .error(let error):
-            if let socket = client as? WebSocket, let index = sockets.firstIndex(where: { $0 === socket }) {
+            if let index = sockets.firstIndex(where: { $0 === socket }) {
                 sockets.remove(at: index)
             }
             handleError(error)